requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncio-dgram==2.2.0
certifi==2024.7.4
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
loguru==0.7.2
mcstatus==11.1.1
multidict==6.0.5
nonebot-adapter-qq==1.5.0
nonebot2==2.3.2
pydantic==2.8.2
pydantic_core==2.20.1
pygtrie==2.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==72.1.0
sniffio==1.3.1
tomli==2.0.1
typing_extensions==4.12.2
websockets==12.0
win32-setctime==1.1.0
yarl==1.9.4
